+++ backends/cortex_m/TARGETS
@@ -1,0 +1,21 @@
+# Copyright 2026 Arm Limited and/or its affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
+
+load("@fbcode_macros//build_defs:python_library.bzl", "python_library")
+
+oncall("executorch")
+
+python_library(
+    name = "quantizer_reporter",
+    srcs = [
+        "quantizer_reporter.py",
+    ],
+    deps = [
+        "//caffe2:torch",
+        "//pytorch/ao:torchao",
+        "fbsource//third-party/pypi/tabulate:tabulate",
+    ],
+)
