+++ backends/cortex_m/quantizer/TARGETS
@@ -17,7 +17,6 @@ python_library(
         "pattern_matcher.py",
         "quantization_configs.py",
         "quantizer.py",
-        "quantizer_reporter.py",
         "quantizer_support.py",
     ],
     deps = [
@@ -27,6 +26,7 @@ python_library(
         "//executorch/backends/arm/quantizer:arm_quantizer_utils",
         "//executorch/backends/arm/quantizer:quantization_annotator",
         "//executorch/backends/arm/quantizer:quantization_config",
+        "//executorch/backends/cortex_m:quantizer_reporter",
         "//pytorch/ao:torchao",
         "fbsource//third-party/pypi/tabulate:tabulate",
     ],
@@ -42,19 +42,7 @@ python_library(
         "//caffe2:torch",
         "//executorch/backends/arm/quantizer:arm_quantizer_utils",
         "//executorch/backends/arm/quantizer:quantization_config",
+        "//executorch/backends/cortex_m:quantizer_reporter",
         "//pytorch/ao:torchao",
-        ":quantizer_reporter",
-    ],
-)
-
-python_library(
-    name = "quantizer_reporter",
-    srcs = [
-        "quantizer_reporter.py",
-    ],
-    deps = [
-        "//caffe2:torch",
-        "//pytorch/ao:torchao",
-        "fbsource//third-party/pypi/tabulate:tabulate",
     ],
 )
